Remove maid_spec.cr
I'm removing tests for now since I'm not sure how to implement them in a
reasonable way. Most of the programs in src/ work with the file system
directly and in a very specific manner.

In the future I may write tests for the programs in src/ if they ever
become used enough to warrant a formal specification.