Remove maid
As much as maid was fun to write (and use!), having two separate files
for each of your dotfiles is a fundamentally flawed design. Using GNU
Stow for dotfiles is a much better way to get things done since the
files are linked: updating ~/.vimrc will automatically update the .vimrc
in your dotfiles directory, and vice versa.

Additionally, GNU Stow makes it easy to pick and choose the dotfiles
that you want to use. If you don't want to use a certain configuration
anymore, it's easy to "stow -D <config>" instead of having to manually
go into the .config directory and remove the offending files.