Move zsh files to fish directory
In reality, zsh is not needed here and shouldn't be supported. Many
defaults of fish require hacky solutions to work in zsh; the investment
and maintenance needed is not worth it.

Eventually, zsh will be removed completely. This is the first step
towards accomplishing that goal.

Unused files and other commands will be simplified in attempt to make my
dotfiles more minimal and use the defaults, which helps a lot when using
a shell without my settings.