Update ranger config
As much as I *not* want to use ranger, it is still very much useful for
easily seeing the structure of your files at a glance. To make
maintenance simpler for me, I removed some of the config settings I
never used and removed image previews since I can just use a proper
desktop environment for that.