Remove waterfox config
Now that I've gone through the stage of trying various browser
derivatives, I've migrated to Firefox and Chromium again, primarily due
to the lack of consistent updates, and partially due to the need to
check upstream a bit more for reported issues. In other words, using
what's in the official repos saves me a lot of time and hassle.