Update Plasma with bspwm config
Since Plasma does not respect feh, it makes no sense to set a
background for it.

Note that these settings apply to Plasma with bspwm, *not* bspwm with
KDE settings. You should not have to use Plasma with bspwm, since bspwm
with KDE settings should do everything you want and much more.

In the future, I may remove Plasma with bspwm support altogether.