Make post-wal commands conditional
This commit makes the post-wal script a bit more aware of its
environment. This is only to support Plasma with bspwm, which
needs the bspwm settings but not the dunst settings.

Since bspwm with KDE settings performs exceptionally well, I may drop
Plasma with bspwm support altogether and revert this commit at a later
date. The overhead of wmctrl and grep may not be worth it.

Additionally, the filename should probably be changed to a more
generic one in order to support more software as needed.