Use symlinks with pywal templates
Instead of manually changing every instance where a program could run to
use the config in the cache directory, and instead of copying files
every time pywal is updated, it makes sense to create symbolic links for
these files instead.

This ensures that the right settings are used even without knowing the
passed parameters.