Fix termite padding to support changing color schemes
The solution I used previously for padding in termite did not change the
padding color when other software such as pywal changed the color scheme.

This fixes that, so now it is possible to change color schemes
fully in termite without having to spawn a new terminal. This is
very nice, and also works well with termite's transparency feature.