stow: Add directories for code, desktop, and editorconfig
The desktop directory is used for desktop-related things that may be
useful outside of bspwm (such as caps lock functioning as ctrl +
escape). Whether or not this is needed is questionable, but it is an
option for now.

It may be better to further split these directories and avoid the
generic word "desktop" altogether.