fedora: Update Makefile
This hasn't been tested yet, but it *should* work.

Most of the software I install is handled by the RPM, with the
exception of software not present in the official repositories.

The install scripts for these software are stored locally in this
repository since I want an easy way to keep track of the diffs. Of
course, you should run the `make update` command yourself to get the
latest updates and visually compare any differences from upstream.