fedora: Add vendor scripts
These scripts can be manually updated by running the `make update`
command. When you do so, visually compare the diffs. This is to
ensure that nothing particularly naughty happened upstream.

Note that this is not a fool-proof solution since one can simply
modify the content returned by any other location a script requests.