rofi: Scale windows pragmatically with launch.sh
Instead of manually setting the DPI to 192 or other values, we can
just take the value from xrdb instead.

This script, in combination with GTK's settings.ini if you are in a
non-GNOME environment, is everything you need to launch both GTK and
Qt applications at the appropriate DPI, with their respective themes
applied as well.