sxhkd: Add bspc hidden flag functionality
Note that window order is not based on the order in which windows are
hidden, but something else entirely. For this reason, it is recommended
to hide only one window at once.

A more robust solution would be to let the user choose from a list of
available hidden windows, potentially through rofi's dmenu capability.