meta: Change directory name from dotfiles to .files
Besides the fact that .files is indisputably cooler, using a literal dot
instead of the actual word makes this a hidden directory, which is great
since configuration files are expected to be hidden by default.

This also keeps our home dir nice and tidy, with only directories
generated by xdg-user-dirs.