meta: Add git branch information to fish prompt
If I stayed on zsh instead of switch to fish, I would've probably never
known how easy it is to add certain information to the prompt. Since
the prompt itself is just a function, you can run any commands you want
inside it to get information, including git commands.

This commit adds the current branch you're on only when inside a git
repository, and only when you're not in a tty.