PKGBUILDs: Merge tari-desktop with tari-core
Since the collective system is known as Tari, having individual parts
for the pieces that work together to make the whole is not ideal.

Now that I understand more about how Arch Linux works, if I ever needed
to create a non-Tari operating system, I would just use pacman to install
packages as needed, or modify this PKGBUILD to create a new meta package
specific to that system.