compton: Add custom shadow settings
With the default shadow settings, gapless windows would have a shadow
extending far into the polybar above.

This change makes gapless windows show a light separator shadow that
distinguishes the window from the bar. It also fixes a problem with
the appearance of the dock shadow in less noisy environments, while
maintaining the shadow look for floating and tiled windows at the
same time.