polybar: Switch label-mode with label-state
Now that the desktop labels cover less width and have lower padding,
it makes sense to place label-mode on the right side instead of the left.

This fixes an issue with padding I previously encountered with the old
desktop labels, and also prevents the desktop labels from being shifted
to the right when a label-mode is present.