docs: Clean up README
Since most of the README wouldn't be relevant to most users anyway,
I've gone ahead and removed it. In the future I may consider writing
a brief guide on how to set up certain things, but for now I'm focusing
more on the dotfiles and bootstrap aspect itself instead of trying to
treat everything as a collective whole.