Remove tari-core meta package
Having a meta package was cool and it got the job done, but it's
inconvenient to use when adding and removing a lot of packages.

One alternative I'm looking at is simply keeping track of all the
explicitly installed packages and storing that in a text file. This
makes it easy to keep track of all the installed packages without
introducing downsides, and new machines that don't need certain
packages can simply delete those lines.