Remove yay stuff from bootstrap script
Note that I need to update the script later or simply provide a series
of instructions so everything isn't dependant on one script.

In the past it was nice to install Arch Linux with minimal if any
console intervention, but the practicality of this is questionable
since you only have to install Arch once. A more specialized script
could be useful for mass installations, although in this case I assume
one would create such a script on-the-spot.