meta: Remove XDG_CURRENT_DESKTOP=KDE
This was cool when we had KDE-specific applications, but since I'm
prioritizing ranger and nautilus now, dolphin isn't needed. Since I'm
focusing heavily on terminal and web-based applications, there is less
need to customize KDE applications specifically.

Two other advantages to this is that I no longer have to manually update
the colors in kdeglobals, and most if not all of the environment can be
programmatically set up with minimal effort.