bspwm: Use a more interesting desktop layout
Since I usually have a browser and terminal emulator open most of the
time, I have placed them back as desktop 1 and 2 by default. I'm used to
the file browser being in 3, and 4 serves as media, which is important
for language immersion in particular.

The other 6 icons are numbers for individuals that know how to read
other languages.