kitty: Enable dynamic background opacity
At some point I disabled this setting, possibly because of a bug or
other issue at the time, most likely related to my use of w3m or a
similar image preview script.

Now that I take advantage of all of kitty's features in 2022, especially
since it has proper fcitx support, there shouldn't be a reason to not
enable dynamic background opacity, as it seems to work flawlessly.