meta: Remove code
I've been trying to like vscode for years, especially since it was
widely regarded as one of the best editors of all time. As cool as VS
Code is, however, I always ended up using neovim and a terminal since
it's simply easier for me to get things done that way.