meta: Remove alacritty
Although alacritty is an amazing terminal emulator, I am not interested
in maintaining configuration files for it at this time.

The current config works and is an excellent replica of the current
kitty config. If you have a desire to use alacritty, you can use the
files in this commit to get amazing support for things like pywal.

The reason alacritty is being removed is because it doesn't fully
support fcitx. Currently "inline input" doesn't work in Xorg, and it's
impossible to see input under GNOME Wayland.

For simplicity, only the terminal emulator that works the best for me
will be officially supported, since I've never had a reason to use
alacritty anyway.

Source: https://github.com/alacritty/alacritty/issues/1613