meta: Add librewolf
This was my configuration for librewolf, a web browser I used for a few
months before ultimately deciding to use firefox again for simplicity.

Although librewolf is useful for giving individuals immediate access to
a solid browser with ublock origin pre-installed, the downsides such as
not having automatic updates for users that need it the most, as well as
constantly depending on another project to update in a timely manner,
do not seem worth it in the long-term.