meta: Add dual-function-keys
dual-function-keys is an amazing piece of software that solves many if
not all of the long-lasting issues I had with binding caps lock to ctrl
and escape. I have additionally configured it in a way such that print
screen can double as a right super key, due to it being next to right
alt and right ctrl on the T14.