meta: Remove bspwm
After over 5 years of bspwm, I have decided to enjoy myself in the
luxurious life that is GNOME.

Using bspwm and window managers in general was an invaluable learning
experience that gave me a deep understanding of many of the novelties
of the current linux desktop computing model. It had a profound impact
on my understanding of how operating systems work in general, and I
now wish to move on and enjoy modern GNOME simplicity.