meta: Remove tint2
tint2 is a great piece of software that I enjoyed playing with, however
it is drastically easier and more convenient to simply use dash-to-panel
in GNOME if you're looking for that traditional taskbar-like experience.

As a bonus, such a taskbar would work under both Xorg and Wayland, have
features such as preview on hover, and won't have anti-features such as
the bar not being clickable unless you perform Xorg shenanigans.