Change workspace animation to vertical
Vertical is nice since the animation is faster and covers less of the
screen. It works well with the master layout since you can easily see
the master window of each workspace.

Being able to have unique partitioning schemes for each workspace (as
long as they're using the master layout) is a nice advantage of Hyprland
over other compositors like river.