meta: Switch from ranger to joshuto
Joshuto is *significantly* faster than ranger and is written in Rust
instead of Python. Although both ranger and joshuto have not seen a new
release in a while, the future of joshuto seems more promising.

Joshuto is additionally faster than lf and, similar to lf, does not hang
when previewing images with kitty.