specializations: Add plasma
This *works*, and the best part is I didn't have to do *anything* (besides
write this configuration file, that is).

Thanks to NixOS, it is possible to have GNOME, Plasma, Hyprland, and
whatever else you want installed on the same computer without those
desktop environments conflicting with each other. This configuration is
done in a fully reproducible and declarative setup with minimal code,
without having to modify any external files or run any imperative
commands.