meta: Replace absolute workspaces with relative workspaces
Instead of remembering which workspace an application is in, it's easier
to simply go to the previous or next workspace until reaching the desired
application.

This has the advantage of less keybinds used and no longer having to reach
across the keyboard when dealing with 6 or more workspaces.

Workspace state can be handled by ironbar's launcher instead, which has
the additional benefit of guaranteeing that you see all open applications.

hyprland-relative-workspace is used here for a GNOME-like workspace
experience. Hyprland's built-in m+1/m-1 would cycle the workspaces
instead of opening an empty one, and the recently merged r+1/r-1 does
not skip empty workspaces in-between other workspaces.