chore: Drop unused packages
I only used inkscape to convert svg to png with the following command:
`fd -e svg -x inkscape -w 512 -h 512 "{}" -o "{.}.png"`

gthumb was crashing and I'd rather not have to deal with a non-trivial
GUI program that crashes on launch.

nb and obsidian are cute but ultimately I found quartz to be better for
my use case.