overlays(kitty): Patch wlroots copying bug
This fixes an issue where copying text would break until changing
workspaces and consequently windows.

This *was* fixed in 0.32.0 as a result of [1], however kitty 0.33.1 with
the recent fontconfig workaround patch[2] ends up showing CJK characters
with a very thin weight, causing them to be inconsistent relative to the
surrounding text and more difficult to read.

[1]: https://github.com/kovidgoyal/kitty/issues/6890
[2]: https://github.com/kovidgoyal/kitty/commit/d2c21ee2973ecda23825b8942989c11caa8193ff